add shield, add chest armour as extension of Armour rather than just using armour as it's own object, add images to all armour files, weapons and shields, revamp activity to display all character stats and equiped items seperately so that images can be added, update CharacterActivity to include new textViews and ImageViews, make changes to toString as necessary, make changes to player to reflect shield addition and Armour changes, add new player stats, accuracy, speed, reactions, intelligence

+++ app/src/main/java/com/example/medievalrebuild/Enemies/BossEnemy.java
@@ -6,10 +6,10 @@ public class BossEnemy extends Enemy implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    int reduceDurability;
+    double reduceDurability;
 
 
-    public BossEnemy(String name, int health, int damage, boolean original, int imageId, int reduceDurability) {
+    public BossEnemy(String name, double health, double damage, boolean original, int imageId, double reduceDurability) {
 
         super(name, health, damage, original, imageId);
 
@@ -17,7 +17,7 @@ public class BossEnemy extends Enemy implements Serializable {
 
     }
 
-    public int getReduceDurability(){
+    public double getReduceDurability(){
 
         return this.reduceDurability;
